requirements.txt
Django==5.0.0
djangorestframework==3.14.0
ultralytics==8.0.0
numpy==1.24.3
opencv-python==4.8.0
Pillow==10.0.0
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.2.0
